web/gradebook/infra/admin_bot/requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.1
pillow==11.0.0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
